+++ tests/test_validation.py
@@ -152,6 +152,34 @@ def test_postgres_identifier_length(schema_loader) -> None:
     assert list(validation.run(dataset)) == []  # no validation errors
 
 
+def test_postgres_duplicate_shortnames(schema_loader) -> None:
+    dataset = schema_loader.get_dataset_from_file("duplicate_shortnames.json")
+
+    error = next(validation.run(dataset))
+    assert error
+    assert error.validator_name == "PostgreSQL duplicate shortnames"
+    assert error.message == "Duplicate shortname 'sameName' found for field: 'veld1,veld2'"
+
+
+def test_postgres_duplicate_abbreviated_fieldnames(schema_loader) -> None:
+    dataset = schema_loader.get_dataset_from_file("abbreviated_fieldnames.json")
+
+    error = next(validation.run(dataset))
+    assert error
+    assert error.validator_name == "PostgreSQL duplicate abbreviated fieldnames"
+    assert (
+        error.message
+        == "Fields 'eenVestigingIsGebouwOfEenComplexGebouwenDuurzameUitoefeningActiviteitenOndernemingRechtspersoon',"
+        " 'eenVestigingIsGebouwOfEenComplexGebouwenDuurzameUitoefeningActiviteitenOndernemingRechtspersoon2' share "
+        "the same first 63 characters. Add a shortname."
+    )
+
+
+def test_postgres_duplicate_abbreviated_fieldnames_with_shortname(schema_loader) -> None:
+    dataset = schema_loader.get_dataset_from_file("abbreviated_fieldnames_with_shortname.json")
+    assert list(validation.run(dataset)) == []  # no validation errors
+
+
 def test_identifier_properties(schema_loader) -> None:
     dataset = schema_loader.get_dataset_from_file("identifier_ref.json")
     error = next(validation.run(dataset))